Renamed pylint\typing.py to pylint_typing.py to avoid conflict with Python's built-in typing module
