+++ pylint/checkers/exceptions.py
@@ -19,7 +19,7 @@ from astroid.typing import InferenceResult, SuccessfulInferenceResult
 from pylint import checkers
 from pylint.checkers import utils
 from pylint.interfaces import HIGH, INFERENCE
-from pylint.typing import MessageDefinitionTuple
+from pylint.pylint_typing import MessageDefinitionTuple
 
 if TYPE_CHECKING:
     from pylint.lint import PyLinter
