+++ pylint/checkers/imports.py
@@ -33,7 +33,7 @@ from pylint.exceptions import EmptyReportError
 from pylint.graph import DotBackend, get_cycles
 from pylint.interfaces import HIGH
 from pylint.reporters.ureports.nodes import Paragraph, Section, VerbatimText
-from pylint.typing import MessageDefinitionTuple
+from pylint.pylint_typing import MessageDefinitionTuple
 from pylint.utils import IsortDriver
 from pylint.utils.linterstats import LinterStats
