+++ pylint/constants.py
@@ -12,7 +12,7 @@ import astroid
 import platformdirs
 
 from pylint.__pkginfo__ import __version__
-from pylint.typing import MessageTypesFullName
+from pylint.pylint_typing import MessageTypesFullName
 
 PY310_PLUS = sys.version_info[:2] >= (3, 10)
 PY311_PLUS = sys.version_info[:2] >= (3, 11)
