+++ pylint/lint/pylinter.py
@@ -57,7 +57,7 @@ from pylint.reporters.base_reporter import BaseReporter
 from pylint.reporters.progress_reporters import ProgressReporter
 from pylint.reporters.text import TextReporter
 from pylint.reporters.ureports import nodes as report_nodes
-from pylint.typing import (
+from pylint.pylint_typing import (
     DirectoryNamespaceDict,
     FileItem,
     ManagedMessage,
