+++ pylint/lint/report_functions.py
@@ -10,7 +10,7 @@ from typing import cast
 
 from pylint import checkers, exceptions
 from pylint.reporters.ureports.nodes import Section, Table
-from pylint.typing import MessageTypesFullName
+from pylint.pylint_typing import MessageTypesFullName
 from pylint.utils import LinterStats
