+++ tests/pyreverse/conftest.py
@@ -12,7 +12,7 @@ from astroid.nodes.scoped_nodes import Module
 from pylint.lint import augmented_sys_path, discover_package_path
 from pylint.pyreverse.inspector import Project, project_from_files
 from pylint.testutils.pyreverse import PyreverseConfig
-from pylint.typing import GetProjectCallable
+from pylint.pylint_typing import GetProjectCallable
 
 
 @pytest.fixture()
