+++ tests/pyreverse/test_diagrams.py
@@ -11,7 +11,7 @@ from collections.abc import Sequence
 from pylint.pyreverse.diadefslib import DefaultDiadefGenerator, DiadefsHandler
 from pylint.pyreverse.inspector import Linker
 from pylint.testutils.pyreverse import PyreverseConfig
-from pylint.typing import GetProjectCallable
+from pylint.pylint_typing import GetProjectCallable
 
 
 def test_property_handling(
