+++ tests/pyreverse/test_inspector.py
@@ -19,7 +19,7 @@ import pytest
 from pylint.pyreverse import inspector
 from pylint.pyreverse.inspector import Project
 from pylint.testutils.utils import _test_cwd
-from pylint.typing import GetProjectCallable
+from pylint.pylint_typing import GetProjectCallable
 
 HERE = Path(__file__)
 TESTS = HERE.parent.parent
